// nebiki-photo-judge_broken/src/domain/types.ts
export type AreaId =
  | "hosomaki"
  | "inari"
  | "futomaki_chumaki"
  | "sushi"
  | "onigiri"
  | "sekihan_takikomi"
  | "balance_bento"
  | "chuka_fish"
  | "yakitori"
  | "fry_chicken"
  | "croquette"
  | "tempura"
  | "bento_men";

  export type NextSessionSkipRecord = {
  date: string;
  targetDiscountTime: "18" | "19";
  areaId: AreaId;
};

export type AreaMaster = {
  id: AreaId;
  name: string;
  order: number;
};